@@ -1,27 +1,22 @@
 import 'package:flutter/material.dart';
 
 class CustomColors {
-  static const blackPearl = Color.fromRGBO(30, 39, 46, 1.0);
-  static const londonSquare = Color.fromRGBO(128, 142, 155, 1.0);
-  static const electronBlue = Color.fromRGBO(9, 132, 227, 1.0);
-  static const darkGray = Color.fromRGBO(72, 84, 96, 1.0);
-  static const lightAmericanGray = Color.fromRGBO(99, 110, 114, 1.0);
-  static const greenTeal = Color.fromRGBO(5, 196, 107, 1.0);
-  static const darkMountainGreen = Color.fromRGBO(16, 172, 132, 1.0);
-  static const cashAppGreen = Color.fromRGBO(40, 193, 1, 1.0);
+  ///COMMON COLORS
+  static const carminPink = Color.fromRGBO(235, 77, 75, 1.0);
+  static const turboYellow = Color.fromRGBO(249, 202, 36, 1.0);
   static const facebookBlue = Color.fromRGBO(59, 89, 152, 1.0);
   static const twitterBlue = Color.fromRGBO(29, 161, 242, 1.0);
-  static const clouds = Color.fromRGBO(241, 242, 246, 1.0);
+  static const darkMountainGreen = Color.fromRGBO(16, 172, 132, 1.0);
+  static const cashAppGreen = Color.fromRGBO(40, 193, 1, 1.0);
   static const iosOffWhite = Color(0xFFF9F9F9);
   static const textFieldGray = Color.fromARGB(255, 241, 241, 241);
-  static const transparent = Color(0x00000000);
+  static const steelGray = Color.fromRGBO(61, 61, 61, 1.0);
 
-  //WEBBLEN COLOR SWATCH
+  ///WEBBLEN APP COLOR SWATCH
   static const webblenRed = Color.fromRGBO(253, 36, 61, 1.0);
   static const webblenRedLowOpacity = Color.fromRGBO(253, 36, 61, 0.1);
   static const webblenPink = Color.fromRGBO(248, 87, 166, 1.0);
   static const webblenMatteBlue = Color.fromRGBO(63, 61, 86, 1.0);
-  static const webblenLightGray = Color.fromRGBO(63, 61, 86, 1.0);
   static const webblenMidGray = Color.fromRGBO(182, 182, 189, 1.0);
   static const webblenDarkGray = Color.fromRGBO(25, 25, 25, 1.0);
 }
