@@ -2,36 +2,52 @@
 // Generated file. Do not edit.
 //
 
-import 'package:agora_rtc_engine/agora_rtc_engine_web.dart';
+// ignore_for_file: lines_longer_than_80_chars
+
 import 'package:cloud_firestore_web/cloud_firestore_web.dart';
 import 'package:cloud_functions_web/cloud_functions_web.dart';
-import 'package:connectivity_for_web/connectivity_for_web.dart';
+import 'package:devicelocale/devicelocale_web.dart';
 import 'package:firebase_analytics_web/firebase_analytics_web.dart';
+import 'package:firebase_auth_oauth_web/firebase_auth_oauth_web.dart';
 import 'package:firebase_auth_web/firebase_auth_web.dart';
 import 'package:firebase_core_web/firebase_core_web.dart';
+import 'package:firebase_messaging_web/firebase_messaging_web.dart';
 import 'package:firebase_storage_web/firebase_storage_web.dart';
+import 'package:flutter_facebook_auth_web/flutter_facebook_auth_web.dart';
 import 'package:flutter_keyboard_visibility_web/flutter_keyboard_visibility_web.dart';
+import 'package:geolocator_web/geolocator_web.dart';
 import 'package:google_sign_in_web/google_sign_in_web.dart';
+import 'package:image_picker_for_web/image_picker_for_web.dart';
 import 'package:location_web/location_web.dart';
+import 'package:maps_launcher/maps_launcher_web.dart';
+import 'package:shared_preferences_web/shared_preferences_web.dart';
+import 'package:uni_links_web2/uni_links_web.dart';
 import 'package:url_launcher_web/url_launcher_web.dart';
-import 'package:video_player_web/video_player_web.dart';
+import 'package:wakelock_web/wakelock_web.dart';
 
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
 
 // ignore: public_member_api_docs
 void registerPlugins(Registrar registrar) {
-  AgoraRtcEngineWeb.registerWith(registrar);
   FirebaseFirestoreWeb.registerWith(registrar);
   FirebaseFunctionsWeb.registerWith(registrar);
-  ConnectivityPlugin.registerWith(registrar);
+  DevicelocaleWeb.registerWith(registrar);
   FirebaseAnalyticsWeb.registerWith(registrar);
+  FirebaseAuthOAuthWeb.registerWith(registrar);
   FirebaseAuthWeb.registerWith(registrar);
   FirebaseCoreWeb.registerWith(registrar);
+  FirebaseMessagingWeb.registerWith(registrar);
   FirebaseStorageWeb.registerWith(registrar);
+  FlutterFacebookAuthPlugin.registerWith(registrar);
   FlutterKeyboardVisibilityPlugin.registerWith(registrar);
+  GeolocatorPlugin.registerWith(registrar);
   GoogleSignInPlugin.registerWith(registrar);
+  ImagePickerPlugin.registerWith(registrar);
   LocationWebPlugin.registerWith(registrar);
+  MapsLauncherWeb.registerWith(registrar);
+  SharedPreferencesPlugin.registerWith(registrar);
+  UniLinksPlugin.registerWith(registrar);
   UrlLauncherPlugin.registerWith(registrar);
-  VideoPlayerPlugin.registerWith(registrar);
+  WakelockWeb.registerWith(registrar);
   registrar.registerMessageHandler();
 }
