@@ -1,63 +1,115 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_core/firebase_core.dart';
+import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:provider/provider.dart';
-import 'package:webblen/home_page.dart';
-import 'package:webblen/pages/auth_pages/login_page.dart';
-import 'package:webblen/pages/onboarding/onboarding_path_select_page.dart';
-import 'package:webblen/root_page.dart';
-import 'package:webblen/styles/flat_colors.dart';
+import 'package:flutter_local_notifications/flutter_local_notifications.dart';
+import 'package:stacked_services/stacked_services.dart';
+import 'package:stacked_themes/stacked_themes.dart';
+import 'package:webblen/app/app.locator.dart';
+import 'package:webblen/app/app.router.dart';
+import 'package:webblen/app/theme_config.dart';
+import 'package:webblen/services/firestore/common/firebase_messaging_service.dart';
+import 'package:webblen/services/firestore/data/user_data_service.dart';
+import 'package:webblen/services/reactive/user/reactive_user_service.dart';
+import 'package:webblen/ui/bottom_sheets/setup_bottom_sheet_ui.dart';
 
-import 'pages/onboarding/onboarding_complete_page.dart';
-import 'pages/user_pages/setup_page.dart';
+import 'models/webblen_user.dart';
 
-void main() {
-  runApp(WebblenApp());
+Future<void> _firebaseMessagingBackgroundHandler(RemoteMessage message) async {
+  // If you're going to use other Firebase services in the background, such as Firestore,
+  // make sure you call `initializeApp` before using other Firebase services.
+  await Firebase.initializeApp();
+  print('Handling a background message ${message.messageId}');
+  FirebaseMessagingService().configureFirebaseMessagingListener();
+}
+
+/// Create a [AndroidNotificationChannel] for heads up notifications
+const AndroidNotificationChannel channel = AndroidNotificationChannel(
+  'high_importance_channel', // id
+  'High Importance Notifications', // title
+  'This channel is used for important notifications.', // description
+  importance: Importance.high,
+);
+
+/// Initialize the [FlutterLocalNotificationsPlugin] package.
+final FlutterLocalNotificationsPlugin flutterLocalNotificationsPlugin = FlutterLocalNotificationsPlugin();
+
+void main() async {
+  // Register all the models and services before the app starts
+  await ThemeManager.initialise();
+
+  WidgetsFlutterBinding.ensureInitialized();
+  await Firebase.initializeApp();
+
+  setupLocator();
+  setupBottomSheetUI();
+  setupSnackBarUi();
+
+  // Set the background messaging handler early on, as a named top-level function
+  FirebaseMessaging.onBackgroundMessage(_firebaseMessagingBackgroundHandler);
+  if (flutterLocalNotificationsPlugin.resolvePlatformSpecificImplementation<AndroidFlutterLocalNotificationsPlugin>() != null) {
+    flutterLocalNotificationsPlugin.resolvePlatformSpecificImplementation<AndroidFlutterLocalNotificationsPlugin>()!.createNotificationChannel(channel);
+  }
+
+  // Update the iOS foreground notification presentation options to allow
+  // heads up notifications.
+  await FirebaseMessaging.instance.setForegroundNotificationPresentationOptions(
+    alert: true,
+    badge: true,
+    sound: false,
+  );
+
+  //Get User Instance if Previously Logged In
+  await setupAuthListener();
+
+  //Set Device Orientation
+  SystemChrome.setPreferredOrientations([DeviceOrientation.portraitUp]).then((_) {
+    runApp(WebblenApp());
+  });
+}
+
+///AUTH & NOTIFICATIONS
+Future<void> setupAuthListener() async {
+  ReactiveUserService _reactiveUserService = locator<ReactiveUserService>();
+  UserDataService _userDataService = locator<UserDataService>();
+  FirebaseAuth.instance.authStateChanges().listen((event) async {
+    if (event != null) {
+      _reactiveUserService.updateUserLoggedIn(true);
+      WebblenUser user = await _userDataService.getWebblenUserByID(event.uid);
+      _reactiveUserService.updateUser(user);
+    }
+  });
+}
+
+///SNACK BAR
+void setupSnackBarUi() {
+  final service = locator<SnackbarService>();
+  service.registerSnackbarConfig(
+    SnackbarConfig(
+      backgroundColor: Colors.red,
+      textColor: Colors.white,
+      mainButtonTextColor: Colors.black,
+    ),
+  );
 }
 
 class WebblenApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    SystemChrome.setPreferredOrientations([
-      DeviceOrientation.portraitUp,
-      DeviceOrientation.portraitDown,
-    ]);
-    return FutureBuilder(
-        future: Firebase.initializeApp(),
-        builder: (context, snapshot) {
-          if (snapshot.hasError) {
-            return Container();
-          }
-          if (snapshot.connectionState == ConnectionState.done) {
-            return MultiProvider(
-              providers: [
-                StreamProvider<User>.value(value: FirebaseAuth.instance.authStateChanges()),
-              ],
-              child: MaterialApp(
-                title: 'Webblen',
-                debugShowCheckedModeBanner: false,
-                theme: ThemeData(
-                  brightness: Brightness.light,
-                  primaryColor: FlatColors.webblenRed,
-                  accentColor: FlatColors.darkGray,
-                  textTheme: Theme.of(context).textTheme.apply(
-                        fontFamily: "Helvetica Neue",
-                      ),
-                ),
-                home: RootPage(),
-                routes: <String, WidgetBuilder>{
-                  '/root': (BuildContext context) => RootPage(),
-                  '/login': (BuildContext context) => LoginPage(),
-                  '/setup': (BuildContext context) => SetupPage(),
-                  '/onboarding': (BuildContext context) => OnboardingPathSelectPage(),
-                  '/home': (BuildContext context) => HomePage(),
-                  '/onboarding_complete': (BuildContext context) => OnboardingCompletePage(),
-                },
-              ),
-            );
-          }
-          return Container(color: Colors.white);
-        });
+    return ThemeBuilder(
+      lightTheme: regularTheme,
+      darkTheme: darkTheme,
+      builder: (context, regularTheme, darkTheme, themeMode) => MaterialApp(
+        debugShowCheckedModeBanner: false,
+        title: 'Webblen',
+        theme: regularTheme,
+        darkTheme: darkTheme,
+        themeMode: themeMode == ThemeMode.system ? ThemeMode.light : themeMode,
+        initialRoute: Routes.RootViewRoute,
+        onGenerateRoute: StackedRouter().onGenerateRoute,
+        navigatorKey: StackedService.navigatorKey,
+      ),
+    );
   }
 }
